+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.8.0] - 2026-09-07
 
 ### Security
 - The actions endpoint enforces CSRF itself with `protect_from_forgery`